Move all code inside the 'mutual_debt' package
