+++ mutual_debt/mutual_debt_simplification.py
@@ -1,6 +1,7 @@
-from Mutual_Debt_Simplification.data_structures import Graph
+from mutual_debt.data_structures import Graph
 
 MIN_DISPLAY_VALUE = 0.01  # debts smaller than this value will be ommited from generated graphs
+
 
 # # # Input # # #
 def debt_list_to_graph(debt_list: list, names: dict) -> Graph:
