+++ mutual_debt/tests/test_mutual_debt_simplification.py
@@ -1,5 +1,7 @@
 import unittest
-from mutual_debt_simplification import simplify_debt_graph, debt_list_to_graph
+
+from mutual_debt.mutual_debt_simplification import simplify_debt_graph, \
+    debt_list_to_graph
 
 test_names = {'A': 'A', 'B': 'B', 'C': 'C', 'D': 'D', 'E': 'E'}
